requirements.txt
transformers
pytorch
pandas
gdown